test: give the status layers one home each, and attachments their own file
The status feature file walked the field through every layer, so a repository
count or a WebSocket command had two places it could be tested. The layer
files are now the home: the repository's counts and status index sit beside
the vocabulary mutators that reindex them, and item/create, item/update,
item/list, items/bulk and stats sit beside status/create and status/delete.
The feature file keeps what decides the field's meaning — the model, the
migration, the stored payload, import/export, the service schemas and the
live status set.

The attachment retitle and reorder tests were in the status file with nothing
in common with it; they move to a file named for what they cover.

No test is dropped here: the offline suite collects the same count before and
after.

Refs #353
